// src/generated/pokemon_pb.ts
// @generated by protoc-gen-es v2.2.3 with parameter "target=ts"
// @generated from file pokemon.proto (package pokemon.v1, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage, GenService } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc, serviceDesc } from "@bufbuild/protobuf/codegenv1";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file pokemon.proto.
 */
export const file_pokemon: GenFile = /*@__PURE__*/
  fileDesc("Cg1wb2tlbW9uLnByb3RvEgpwb2tlbW9uLnYxIhwKDlBva2Vtb25SZXF1ZXN0EgoKAmlkGAEgASgFIocBCg9Qb2tlbW9uUmVzcG9uc2USCgoCaWQYASABKAUSDAoEbmFtZRgCIAEoCRIXCg9iYXNlX2V4cGVyaWVuY2UYAyABKAUSDgoGaGVpZ2h0GAQgASgFEhIKCmlzX2RlZmF1bHQYBSABKAgSDQoFb3JkZXIYBiABKAUSDgoGd2VpZ2h0GAcgASgFMlsKDlBva2Vtb25TZXJ2aWNlEkkKDEZldGNoUG9rZW1vbhIaLnBva2Vtb24udjEuUG9rZW1vblJlcXVlc3QaGy5wb2tlbW9uLnYxLlBva2Vtb25SZXNwb25zZSIAQiVaI2dvX2Nvbm5lY3QvZ2VuL3Bva2Vtb24vdjE7cG9rZW1vbnYxYgZwcm90bzM");

/**
 * @generated from message pokemon.v1.PokemonRequest
 */
export type PokemonRequest = Message<"pokemon.v1.PokemonRequest"> & {
  /**
   * @generated from field: int32 id = 1;
   */
  id: number;
};

/**
 * Describes the message pokemon.v1.PokemonRequest.
 * Use `create(PokemonRequestSchema)` to create a new message.
 */
export const PokemonRequestSchema: GenMessage<PokemonRequest> = /*@__PURE__*/
  messageDesc(file_pokemon, 0);

/**
 * @generated from message pokemon.v1.PokemonResponse
 */
export type PokemonResponse = Message<"pokemon.v1.PokemonResponse"> & {
  /**
   * @generated from field: int32 id = 1;
   */
  id: number;

  /**
   * @generated from field: string name = 2;
   */
  name: string;

  /**
   * @generated from field: int32 base_experience = 3;
   */
  baseExperience: number;

  /**
   * @generated from field: int32 height = 4;
   */
  height: number;

  /**
   * @generated from field: bool is_default = 5;
   */
  isDefault: boolean;

  /**
   * @generated from field: int32 order = 6;
   */
  order: number;

  /**
   * @generated from field: int32 weight = 7;
   */
  weight: number;
};

/**
 * Describes the message pokemon.v1.PokemonResponse.
 * Use `create(PokemonResponseSchema)` to create a new message.
 */
export const PokemonResponseSchema: GenMessage<PokemonResponse> = /*@__PURE__*/
  messageDesc(file_pokemon, 1);

/**
 * @generated from service pokemon.v1.PokemonService
 */
export const PokemonService: GenService<{
  /**
   * @generated from rpc pokemon.v1.PokemonService.FetchPokemon
   */
  fetchPokemon: {
    methodKind: "unary";
    input: typeof PokemonRequestSchema;
    output: typeof PokemonResponseSchema;
  },
}> = /*@__PURE__*/
  serviceDesc(file_pokemon, 0);